Create AutoRefreshIdentity.ahk
Quick AutoHotkey script to refresh the identity in Tor Browser until you find an exit node that is not blacklisted on the website that you are trying to access.

To use, simply execute the script then focus the TorBrowser window sitting at the "Access Denied" or "Forbidden" page of the website that you are trying to visit. The window will need to remain focused.

Title matching can be tweaked in line 12, timeout on line 15.